+++ Implementations/IRootsFindAlg.h
@@ -5,6 +5,10 @@
 class IRootsFindAlg
 {
 public:
+	virtual ~IRootsFindAlg()
+	{
+
+	};
 	/// <summary>
 	/// find the roots for a given polynomial
 	/// </summary>
@@ -13,7 +17,7 @@ public:
 	/// <param name="roots">array for the results, with size = degree</param>
 	/// <returns>the number of real roots found</returns>
 	virtual int findRoots(double* coefficients, int degree, double* roots) = 0;
-
+	virtual void init(int degrre) = 0;
 
 };
 #endif //SHIELD_IRootsFindAlg_H