+++ TestedOBCApp/Main/Main.cpp
@@ -4,24 +4,9 @@
 
 #include "Version.h"
 #include "WinTimer.h"
+#include "Factory.h"
+#include "TestManager.h"
 #include "MainProcess.h"
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 int main() {
     std::cout << "Version " 
@@ -29,8 +14,9 @@ int main() {
         << ProjectVersions::VERSION_MINOR << "." 
         << ProjectVersions::VERSION_PATCH << std::endl;
 
-    WinTimer timer;
-    MainProcess mainProcess(timer);
+    Factory *factory = Factory::getReferance();
+    TestManager testManager(factory->getTimer());
+    MainProcess mainProcess(&testManager);
     mainProcess.process();
     
     return 0;